Refactor example to use common components
