+++ website/docs/recommended-patterns.mdx
@@ -24,7 +24,7 @@ your form elements.
 
 For example, you may define your own `TextField` field as:
 
-import TextField from "!!raw-loader!../examples/src/recommended-patterns/avoid-using-native-form-elements-directly/TextField";
+import TextField from "!!raw-loader!../examples/src/common/TextField";
 
 <CodeBlock className="language-tsx">{TextField}</CodeBlock>
