+++ website/examples/src/recommended-patterns/define-a-shared-type-for-field-issues/TextField.tsx
@@ -17,11 +17,11 @@ export const TextField = (props: Props) => {
         onBlur={props.onBlur}
         disabled={props.disabled}
       />
-      <p>
+      <ul>
         {props.issues?.map((issue) => (
-          <span>{issue.message}</span>
+          <li key={issue.message}>{issue}</li>
         ))}
-      </p>
+      </ul>
     </div>
   );
 };
